+++ src/backend/distributed/operations/citus_split_shard_by_split_points.c
@@ -77,6 +77,7 @@ LookupSplitMode(Oid shardSplitModeOid)
 	Datum enumLabelDatum = DirectFunctionCall1(enum_out, shardSplitModeOid);
 	char *enumLabel = DatumGetCString(enumLabelDatum);
 
+	/* Extend with other modes as we support them */
 	if (strncmp(enumLabel, "blocking", NAMEDATALEN) == 0)
 	{
 		shardSplitMode = BLOCKING_SPLIT;
@@ -88,7 +89,8 @@ LookupSplitMode(Oid shardSplitModeOid)
 	/* Extend with other modes as we support them */
 	else
 	{
-		ereport(ERROR, (errmsg("Invalid label for enum: %s", enumLabel)));
+		ereport(ERROR, (errmsg("Invalid split mode: %s. Expected split mode is blocking.",
+							   enumLabel)));
 	}
 
 	return shardSplitMode;