+++ src/backend/distributed/operations/worker_split_copy_udf.c
@@ -81,13 +81,18 @@ worker_split_copy(PG_FUNCTION_ARGS)
 																				 shardIntervalToSplitCopy,
 																				 splitCopyInfoList);
 
-	char *sourceShardToCopyName = generate_qualified_relation_name(
+	Oid sourceShardToCopySchemaOId = get_rel_namespace(
 		shardIntervalToSplitCopy->relationId);
+	char *sourceShardToCopySchemaName = get_namespace_name(sourceShardToCopySchemaOId);
+	char *sourceShardToCopyName = get_rel_name(shardIntervalToSplitCopy->relationId);
 	AppendShardIdToName(&sourceShardToCopyName, shardIdToSplitCopy);
+	char *sourceShardToCopyQualifiedName = quote_qualified_identifier(
+		sourceShardToCopySchemaName,
+		sourceShardToCopyName);
 
 	StringInfo selectShardQueryForCopy = makeStringInfo();
 	appendStringInfo(selectShardQueryForCopy,
-					 "SELECT * FROM %s;", sourceShardToCopyName);
+					 "SELECT * FROM %s;", sourceShardToCopyQualifiedName);
 
 	ParamListInfo params = NULL;
 	ExecuteQueryStringIntoDestReceiver(selectShardQueryForCopy->data, params,
@@ -197,9 +202,9 @@ CreateShardCopyDestReceivers(EState *estate, ShardInterval *shardIntervalToSplit
 	char *sourceShardNamePrefix = get_rel_name(shardIntervalToSplitCopy->relationId);
 	foreach_ptr(splitCopyInfo, splitCopyInfoList)
 	{
-		char *destinationShardSchemaName = get_namespace_name(get_rel_namespace(
-																  shardIntervalToSplitCopy
-																  ->relationId));
+		Oid destinationShardSchemaOid = get_rel_namespace(
+			shardIntervalToSplitCopy->relationId);
+		char *destinationShardSchemaName = get_namespace_name(destinationShardSchemaOid);
 		char *destinationShardNameCopy = pstrdup(sourceShardNamePrefix);
 		AppendShardIdToName(&destinationShardNameCopy, splitCopyInfo->destinationShardId);