+++ src/backend/distributed/sql/downgrades/citus--11.0-2--11.0-1.sql
@@ -1,5 +1,6 @@
 #include "../udfs/citus_shards_on_worker/11.0-1.sql"
 #include "../udfs/citus_shard_indexes_on_worker/11.0-1.sql"
+#include "../udfs/citus_finalize_upgrade_to_citus11/11.0-1.sql"
 
 DROP FUNCTION  pg_catalog.citus_disable_node(text, integer, bool);
 CREATE FUNCTION pg_catalog.citus_disable_node(nodename text, nodeport integer, force bool default false)
@@ -15,5 +16,4 @@ DROP FUNCTION pg_catalog.citus_is_coordinator();
 DROP FUNCTION pg_catalog.run_command_on_coordinator(text,boolean);
 
 DROP FUNCTION pg_catalog.start_metadata_sync_to_all_nodes();
-DROP FUNCTION pg_catalog.citus_finalize_upgrade_to_citus11(boolean);
 DROP PROCEDURE pg_catalog.citus_finish_citus_upgrade();