+++ src/backend/distributed/sql/downgrades/citus--11.1-1--11.0-3.sql
@@ -46,8 +46,6 @@ CREATE FUNCTION pg_catalog.worker_repartition_cleanup(bigint)
  STRICT
 AS 'MODULE_PATHNAME', $function$worker_repartition_cleanup$function$;
 
-DROP TYPE IF EXISTS citus.split_mode;
-DROP TYPE IF EXISTS citus.split_copy_info;
 DROP FUNCTION pg_catalog.citus_split_shard_by_split_points(
     shard_id bigint,
     split_points text[],
@@ -56,5 +54,12 @@ DROP FUNCTION pg_catalog.citus_split_shard_by_split_points(
 DROP FUNCTION pg_catalog.worker_split_copy(
     source_shard_id bigint,
     splitCopyInfos citus.split_copy_info[]);
+DROP TYPE citus.split_mode;
+DROP TYPE citus.split_copy_info;
 
-#include "../../../columnar/sql/downgrades/columnar--11.1-1--11.0-2.sql"
+#include "../../../columnar/sql/downgrades/columnar--11.1-1--11.0-3.sql"
+
+DROP FUNCTION pg_catalog.get_all_active_transactions(OUT datid oid, OUT process_id int, OUT initiator_node_identifier int4,
+                                                     OUT worker_query BOOL, OUT transaction_number int8, OUT transaction_stamp timestamptz,
+                                                     OUT global_pid int8);
+#include "../udfs/get_all_active_transactions/11.0-1.sql"