+++ src/test/regress/sql/citus_local_tables_mx.sql
@@ -453,7 +453,17 @@ CREATE VIEW v100 AS SELECT * FROM loc_tb;
 CREATE VIEW v101 AS SELECT * FROM loc_tb JOIN ref_tb USING (a);
 CREATE VIEW v102 AS SELECT * FROM v101;
 
+-- a regular matview that depends on local table
+CREATE MATERIALIZED VIEW matview_101 AS SELECT * from loc_tb;
+-- a matview and a view that depend on the local table + each other
+CREATE VIEW v103 AS SELECT * from loc_tb;
+CREATE MATERIALIZED VIEW matview_102 AS SELECT * from loc_tb JOIN v103 USING (a);
+CREATE OR REPLACE VIEW v103 AS SELECT * from loc_tb JOIN matview_102 USING (a);
+
+SET client_min_messages TO DEBUG1;
+-- auto undistribute
 ALTER TABLE loc_tb ADD CONSTRAINT fkey FOREIGN KEY (a) references ref_tb(a);
+SET client_min_messages TO WARNING;
 
 -- works fine
 select run_command_on_workers($$SELECT count(*) from citus_local_tables_mx.v100, citus_local_tables_mx.v101, citus_local_tables_mx.v102$$);
@@ -463,6 +473,71 @@ ALTER TABLE loc_tb DROP CONSTRAINT fkey;
 select run_command_on_workers($$SELECT count(*) from citus_local_tables_mx.v100$$);
 	select run_command_on_workers($$SELECT count(*) from citus_local_tables_mx.v101$$);
 	select run_command_on_workers($$SELECT count(*) from citus_local_tables_mx.v102$$);
+
+INSERT INTO loc_tb VALUES (1), (2);
+-- test a matview with columnar
+CREATE MATERIALIZED VIEW matview_columnar USING COLUMNAR AS SELECT * FROM loc_tb WITH DATA;
+
+-- cant recreate matviews, because the size limit is set to zero, by the GUC
+SET citus.max_matview_size_to_auto_recreate TO 0;
+SELECT citus_add_local_table_to_metadata('loc_tb', true);
+-- remove the limit
+SET citus.max_matview_size_to_auto_recreate TO -1;
+SELECT citus_add_local_table_to_metadata('loc_tb', true);
+
+-- test REFRESH MAT VIEW
+SELECT * FROM matview_101 ORDER BY a;
+REFRESH MATERIALIZED VIEW matview_101;
+SELECT * FROM matview_101 ORDER BY a;
+
+-- verify columnar matview works on a table added to metadata
+SELECT * FROM matview_columnar;
+REFRESH MATERIALIZED VIEW matview_columnar;
+SELECT * FROM matview_columnar ORDER BY a;
+
+-- test with partitioned tables
+SET citus.use_citus_managed_tables TO ON;
+CREATE TABLE parent_1 (a INT UNIQUE) PARTITION BY RANGE(a);
+SET citus.use_citus_managed_tables TO OFF;
+
+CREATE MATERIALIZED VIEW part_matview1 as SELECT count(*) FROM parent_1 JOIN parent_1 p2 ON (true);
+CREATE MATERIALIZED VIEW part_matview2 as SELECT count(*) FROM parent_1 JOIN part_matview1 on (true);
+
+SELECT count(*) FROM citus_local_tables_mx.part_matview1 JOIN citus_local_tables_mx.part_matview2 ON (true);
+
+CREATE TABLE parent_1_child_1 (a int);
+CREATE TABLE parent_1_child_2 (a int);
+
+-- create matviews on partition tables
+CREATE MATERIALIZED VIEW mv1 AS SELECT * FROM parent_1_child_1;
+CREATE MATERIALIZED VIEW mv2 AS SELECT * FROM parent_1_child_2;
+CREATE MATERIALIZED VIEW mv3 AS SELECT parent_1_child_2.* FROM parent_1_child_2 JOIN parent_1_child_1 USING(a);
+CREATE MATERIALIZED VIEW mv4 AS SELECT * FROM mv3;
+
+alter table parent_1 attach partition parent_1_child_1 FOR VALUES FROM (0) TO (10) ;
+
+-- all matviews work
+SELECT count(*) FROM citus_local_tables_mx.mv1;
+SELECT count(*) FROM citus_local_tables_mx.mv2;
+SELECT count(*) FROM citus_local_tables_mx.mv3;
+SELECT count(*) FROM citus_local_tables_mx.mv4;
+
+-- recreate matviews and verify they still work
+alter table parent_1 attach partition parent_1_child_2 FOR VALUES FROM (10) TO (20);
+
+SELECT count(*) FROM citus_local_tables_mx.mv1;
+SELECT count(*) FROM citus_local_tables_mx.mv2;
+SELECT count(*) FROM citus_local_tables_mx.mv3;
+SELECT count(*) FROM citus_local_tables_mx.mv4;
+
+-- verify matviews work after undistributing
+SELECT undistribute_table('parent_1');
+SELECT count(*) FROM citus_local_tables_mx.mv1;
+SELECT count(*) FROM citus_local_tables_mx.mv2;
+SELECT count(*) FROM citus_local_tables_mx.mv3;
+SELECT count(*) FROM citus_local_tables_mx.mv4;
+
+-- todo: add more matview tests once 5968 and 6028 are fixed
 
 -- cleanup at exit
 set client_min_messages to error;